chore: rename `maskitoCompositionValidationPlugin` => `maskitoStrictCompositionPlugin`
