--- projects/demo/src/pages/documentation/plugins/examples/3-composition-validation/mask.ts
@@ -1,6 +1,0 @@
-import {maskitoCompositionValidationPlugin, MaskitoOptions} from '@maskito/core';
-
-export default {
-    mask: /^[0-9０-９]*$/,
-    plugins: [maskitoCompositionValidationPlugin()],
-} as MaskitoOptions;
