+++ projects/demo/src/pages/documentation/plugins/examples/3-strict-composition/mask.ts
@@ -1,0 +1,6 @@
+import {MaskitoOptions, maskitoStrictCompositionPlugin} from '@maskito/core';
+
+export default {
+    mask: /^[0-9０-９]*$/,
+    plugins: [maskitoStrictCompositionPlugin()],
+} as MaskitoOptions;
