+++ projects/demo/src/pages/documentation/plugins/plugins.module.ts
@@ -10,7 +10,7 @@ import {TuiInputModule} from '@taiga-ui/kit';
 import {NextStepsModule} from '../next-steps/next-steps.module';
 import {PluginsDocExample1} from './examples/1-reject/component';
 import {PluginsDocExample2} from './examples/2-initial-calibration/component';
-import {PluginsDocExample3} from './examples/3-composition-validation/component';
+import {PluginsDocExample3} from './examples/3-strict-composition/component';
 import {PluginsDocPageComponent} from './plugins.component';
 
 @NgModule({
